% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.R
\name{binClst_instance}
\alias{binClst_instance}
\title{binClst Instance definition}
\description{
Unless otherwise specified, a \code{binClst} instance refers to any of the binary clustering objects defined in the package, either a \code{binClst} object itself, or any of its child classes, a \code{binClstPath} or a \code{binClstMove} instance.  The latter inherit all slots and functionality defined for the former.
}
