% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.R
\name{binClstPath_instance}
\alias{binClstPath_instance}
\title{binClstPath Instance definition}
\description{
Unless otherwise specified, a \code{binClstPath} instance refers to a \code{binClstPath} object itself, as well as its child class \code{binClstMove}. The latter inherits all slots and functionality defined for the former.
}
