% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{binClstStck-class}
\alias{binClstStck-class}
\title{Binary Clustering Move Class
#
#}
\description{
\code{binClstMove} is a \code{binClstPath} subclass for speed/turn-clustering of \code{Move} objects from the \code{move} R-package. This class inherits all slots and functionality of \code{binClstPath} and \code{Move} objects.
#
}
\details{
#
}
\section{Slots}{

\describe{
\item{\code{bCS}}{A list of either \code{binClstPath} or \code{binClstMove} objects, depending on how the input paths are given.}

\item{\code{bC}}{A \code{binClst} instance with the global speed/turn clustering of the paths in the stack.}
}}

