% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{lkhp}
\alias{lkhp}
\alias{lkhp,binClst-method}
\alias{lkhp,list-method}
\title{Likelihood profile plots}
\usage{
lkhp(obj, offSet = 1)

\S4method{lkhp}{binClst}(obj, offSet = 1)

\S4method{lkhp}{list}(obj, offSet = 1)
}
\arguments{
\item{obj}{A \code{BinClst_instance} or a list of them.}

\item{offSet}{A numeric value indicating an offset to avoid the initial
iterations. This is useful to see the likelihood evolution in the last
iterations where the changes in likelihood are of different order of
magnitude than those at the starting iterations.}
}
\description{
\code{lkhp} likelihood optimization plot.
}
\examples{
# -- apply EMbC to the example path --
mybcp <- stbc(expth)
# -- inspect the likelihood evolution --
lkhp(mybcp)
# -- avoid the initial values --
lkhp(mybcp,10)
}
