% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setc}
\alias{setc}
\title{Sets binClst color palette .}
\usage{
setc(bC, fam = "RdYlBu")
}
\arguments{
\item{bC}{A \link{binClst_instance}.}

\item{fam}{The name of a color family from the Rcolorbrewer R-package,
(default color palette is 'RdYlBu' which is colorblind safe and print
friendly up to 6 colors).}
}
\description{
\code{setc} sets the color palette to a color family from the
  RColorbrewer package
}
\examples{
# -- change the color palette of mybc to "PuOr" --
\dontrun{
setc(mybc,'PuOr')
}
}
