% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{slct}
\alias{slct}
\title{Select a single path from a \code{binClstStck} instance.}
\usage{
slct(stck, pathNmbr)
}
\arguments{
\item{stck}{A \code{binClstStck} instance.}

\item{pathNmbr}{The number of the single path to be selected.}
}
\value{
Returns the single \link{binClstPath_instance} selected.
}
\description{
\code{slct} selects a single path from a \code{binClstStck}
  instance.
}
\examples{
\dontrun{
# -- select path number 3 in mybcpstack --
bcp3 <- slct(mybcpstack,3)
}
}
