% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.R
\docType{data}
\name{x2d}
\alias{x2d}
\title{Synthetic 2D object used in the examples}
\format{See parameter \code{X} of the \link{embc} constructor.}
\description{
An ad-hoc object with a set of bivariate data points synthetically generated by sampling from a four component GMM and their corresponding labels indicating which component of the mixture generated each data point.
}
\keyword{datasets}
