% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{binClstStck-class}
\alias{binClstStck-class}
\title{Binary Clustering Stack Class}
\description{
\code{binClstStck} is a special class for population level speed/turn-clustering of movement trajectories, given either as path data.frames or \code{move} objects.
}
\section{Slots}{

\describe{
\item{\code{bCS}}{A list of either \code{binClstPath} or \code{binClstMove} objects, depending on how the input paths are given.}

\item{\code{bC}}{A \code{binClst} instance with the global speed/turn clustering of the paths in the stack.}
}}

